LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/dirty-contents-reselect-anchor.tentative-expected.txt

Scroll target

FAIL Scroll anchor is re-selected after adjustment if there are dirty descendants at selection time assert_equals: Scroll position should've been preserved expected 1207 but got 0
